test(e2e): run every e2e_flow test against an upgraded chain
setup_test_networks now publishes the deployed v1 bytecode snapshot and
runs the full governance upgrade to current source before handing the
network to the test, so all e2e_flow tests exercise the post-upgrade
state a real network is in (package history {v1, vN}, routing at the
upgraded package id, types defined at the v1 address) instead of a
fresh single-version publish. wait_for_package_convergence moves from
upgrade_tests into upgrade_flow for reuse as the setup barrier.
Fresh-publish coverage remains in the genesis-specific harnesses.
